discord-bot/requirements.txt
discord.py==2.3.2
aio-pika==9.1.4
python-dotenv==1.0.1
